requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
supabase==2.6.0
google-generativeai==0.7.2
pandas==2.2.2
python-dateutil==2.9.0.post0
